+++ src/ft_itoa.c
@@ -6,19 +6,19 @@
 /*   By: nsabbah <nsabbah@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/10 20:43:31 by nsabbah           #+#    #+#             */
-/*   Updated: 2016/11/11 17:26:48 by nsabbah          ###   ########.fr       */
+/*   Updated: 2017/01/10 17:44:12 by nsabbah          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include <stdlib.h>
 
-char	*ft_itoa(int n)
+char	*ft_itoa(long long int n)
 {
 	int				size;
 	int				neg;
 	char			*str;
-	unsigned int	nb;
+	unsigned long long int nb;
 
 	neg = (n < 0) ? 1 : 0;
 	nb = (n < 0) ? -n : n;
